Add unit tests for weekdays string creation
