+++ src/common/utils/date-time/format.test.ts
@@ -1,6 +1,7 @@
 /// <reference types="jest" />
 
-import { formatDateRange } from './format';
+import { Language } from '../../lib/types';
+import { formatDateRange, createWeekdaysStringFromIndices } from './format';
 
 describe('format', () => {
   describe('formatDateRange', () => {
@@ -18,6 +19,45 @@ describe('format', () => {
 
     it('should return string when startDate is missing', () => {
       expect(formatDateRange({ endDate: '2020-12-31' })).toEqual('');
+    });
+  });
+
+  describe('createWeekdaysStringFromIndices', () => {
+    it('should return correct weekday span for the whole week', () => {
+      const weekdayString = createWeekdaysStringFromIndices(
+        [1, 2, 3, 4, 5, 6, 7],
+        Language.FI
+      );
+      expect(weekdayString).toEqual('ma - su');
+    });
+
+    it('should return only single day', () => {
+      const weekdayString = createWeekdaysStringFromIndices([6], Language.FI);
+      expect(weekdayString).toEqual('la');
+    });
+
+    it('should return correct string with individual days and day spans', () => {
+      const weekdayString = createWeekdaysStringFromIndices(
+        [1, 2, 3, 5, 7],
+        Language.FI
+      );
+      expect(weekdayString).toEqual('ma - ke, pe, su');
+    });
+
+    it('should return correct string with multiple day spans', () => {
+      const weekdayString = createWeekdaysStringFromIndices(
+        [1, 2, 3, 5, 6, 7],
+        Language.FI
+      );
+      expect(weekdayString).toEqual('ma - ke, pe - su');
+    });
+
+    it('should return correct weekday span for the whole week with swedish language', () => {
+      const weekdayString = createWeekdaysStringFromIndices(
+        [1, 2, 3, 4, 5, 6, 7],
+        Language.SV
+      );
+      expect(weekdayString).toEqual('Mån - Sön');
     });
   });
 });
